+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.3
+
+* `getReferralDetails()` method is now deprecated use `getReferralInfo()` instead.
+
+* Introduced `onReferralLinkDetected()` in AppLinkListener.
+    * This callback is optional.
+    * It is triggered only when the app is installed and launched for the first time with a referral details.
+
 ## 1.1.2
 
 * Added additional information in Referral Details